move restart action to common actions
