+++ src/state/commonAction.ts
@@ -1,0 +1,5 @@
+export const RESTART_GAME = 'common/restart-game';
+
+export const restartGame = () => ({
+    type: RESTART_GAME
+});
